fix+refactor: SnapshotPaginationProvider race condition
SnapshotPaginationProvider which was acting as both an
ISnapshotPaginationProvider and ISnapshotPagination kept a reference to
a Snapshot internally. This was causing trouble since AsyncFhirService
is a singleton, and ISnapshotPaginationProvider which is a dependency
to AsyncFhirService was therefore also effectively a singleton.

When a race condition between to competing requests occured second
request would overwrite the internal reference of Snapshot in
ISnapshotPaginationProvider before the first request would have time to
call GetPage/GetPageAsync

This was fixed by splitting the implementations of
ISnapshotPaginationProvider and ISnapshotPagination into separate
classes. The implementation of ISnapshotPaginationProvider is now always
returning a freshly instantiated instance of ISnapshotPagination instead
of a reference to itself.
